fix: 헤더 눌렀을 때도 profileDropdown 사라지게 하기
