+++ src/components/default/header/Header.tsx
@@ -1,15 +1,8 @@
 import { useEffect, useState } from 'react';
-import {
-  HeaderContainer,
-  HeaderMenuContainer,
-  HeaderPersonalContainer,
-} from './HeaderStyledComponents';
-import AlarmImg from 'assets/header/header_alarm.svg';
-import ArrowDownImg from 'assets/header/header_arrow_down.svg';
+import { HeaderContainer, HeaderMenuContainer } from './HeaderStyledComponents';
 import { useLocation, useNavigate } from 'react-router-dom';
 import LogoIcon from 'assets/default/icon_logo.svg';
-import ProfileModal from './ProfileModal';
-import { useDropdown } from 'hooks/feature/useDropDown';
+import ProfileDropdown from './ProfileDropdown';
 
 const MENU_ITEM_ARR = ['지원 현황 모아보기', '지원서 추가'];
 
@@ -17,8 +10,6 @@ const Header = () => {
   const { pathname } = useLocation();
   const [isSelect, setIsSelect] = useState('');
   const navigate = useNavigate();
-
-  const { isOpen, toggleDropdown, dropdownRef } = useDropdown();
 
   const handlePage = (item: string) => {
     if (item === '지원서 추가') {
@@ -47,7 +38,7 @@ const Header = () => {
 
   return (
     <HeaderContainer>
-      <div className='headerContainer' ref={dropdownRef}>
+      <div className='headerContainer'>
         <img
           className='headerLogo'
           src={LogoIcon}
@@ -58,27 +49,17 @@ const Header = () => {
         />
         <div className='rightContainer'>
           <HeaderMenuContainer>
-            {MENU_ITEM_ARR.map((e, index) => {
-              return (
-                <div
-                  className={`menuItem ${isSelect === e ? 'active' : ''}`}
-                  key={index}
-                  onClick={() => handlePage(e)}>
-                  {e}
-                </div>
-              );
-            })}
+            {MENU_ITEM_ARR.map((e, index) => (
+              <div
+                className={`menuItem ${isSelect === e ? 'active' : ''}`}
+                key={index}
+                onClick={() => handlePage(e)}>
+                {e}
+              </div>
+            ))}
           </HeaderMenuContainer>
-          <HeaderPersonalContainer>
-            <img src={AlarmImg} className='alarmImg' alt='alarmImg' />
-            <div className='personalBox' onClick={toggleDropdown}>
-              <div className='profile'>고하</div>
-              <div className='profileName'>사용자</div>
-              <img className='arrowDown' alt='arrowDownImg' src={ArrowDownImg} />
-            </div>
-          </HeaderPersonalContainer>
+          <ProfileDropdown />
         </div>
-        {isOpen && <ProfileModal closeModal={toggleDropdown} />}
       </div>
     </HeaderContainer>
   );
